backend-python/ai-service/requirements.txt
# AI服务依赖
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.4.21
Pillow==10.0.0
numpy==1.24.3
